Crate spin_core

Spin core execution engine

This crate provides low-level Wasm functionality required by Spin. Most of this functionality consists of wrappers around wasmtime that narrow the flexibility of wasmtime to the set of features used by Spin (such as only supporting wasmtime’s async calling style).

Re-exports

pub use wasmtime;

Structs

Component

A compiled WebAssembly Component.

Config

Global configuration for EngineBuilder.

Engine

An Engine is a global context for the initialization and execution of Spin components.

EngineBuilder

A builder interface for configuring a new Engine.

Instance

An instantiated component.

InstancePre

A “pre-instantiated” Instance which has all of its arguments already supplied and is ready to instantiate.

Linker

A type used to instantiate Components.

Module

A compiled WebAssembly module, ready to be instantiated.

ModuleInstance

An instantiated WebAssembly module.

State

Host state data associated with individual Stores and Instances.

Store

A Store holds the runtime state of a Spin instance.

StoreBuilder

A builder interface for configuring a new Store.

WasmtimeEngine

An Engine which is a global context for compilation and management of wasm modules.

Enums

Trap

Constants

DEFAULT_EPOCH_TICK_INTERVAL

The default EngineBuilder::epoch_tick_interval.

Traits

AsState

For consumers that need to use a type other than State as the Store data, this trait must be implemented for that type.

Attribute Macros

async_trait