Trait spin_core::ip_name_lookup::Host

pub trait Host {
    // Required methods
    fn resolve_addresses<'life0, 'async_trait>(
        &'life0 mut self,
        network: Network,
        name: String,
        address_family: Option<IpAddressFamily>,
        include_unavailable: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<ResolveAddressStream, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn resolve_next_address<'life0, 'async_trait>(
        &'life0 mut self,
        this: ResolveAddressStream
    ) -> Pin<Box<dyn Future<Output = Result<Result<Option<IpAddress>, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop_resolve_address_stream<'life0, 'async_trait>(
        &'life0 mut self,
        this: ResolveAddressStream
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn non_blocking<'life0, 'async_trait>(
        &'life0 mut self,
        this: ResolveAddressStream
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_non_blocking<'life0, 'async_trait>(
        &'life0 mut self,
        this: ResolveAddressStream,
        value: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 mut self,
        this: ResolveAddressStream
    ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn resolve_addresses<'life0, 'async_trait>( &'life0 mut self, network: Network, name: String, address_family: Option<IpAddressFamily>, include_unavailable: bool ) -> Pin<Box<dyn Future<Output = Result<Result<ResolveAddressStream, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Resolve an internet host name to a list of IP addresses.

See the wasi-socket proposal README.md for a comparison with getaddrinfo.

Parameters:

name: The name to look up. IP addresses are not allowed. Unicode domain names are automatically converted to ASCII using IDNA encoding.

address-family: If provided, limit the results to addresses of this specific address family.

include-unavailable: When set to true, this function will also return addresses of which the runtime thinks (or knows) can’t be connected to at the moment. For example, this will return IPv6 addresses on systems without an active IPv6 interface. Notes:

Even when no public IPv6 interfaces are present or active, names like “localhost” can still resolve to an IPv6 address.

Whatever is “available” or “unavailable” is volatile and can change everytime a network cable is unplugged.

This function never blocks. It either immediately returns successfully with a resolve-address-stream that can be used to (asynchronously) fetch the results. Or it immediately fails whenever name is:

empty

an IP address

a syntactically invalid domain name in another way

References:

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getaddrinfo.html

https://man7.org/linux/man-pages/man3/getaddrinfo.3.html

fn resolve_next_address<'life0, 'async_trait>( &'life0 mut self, this: ResolveAddressStream ) -> Pin<Box<dyn Future<Output = Result<Result<Option<IpAddress>, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns the next address from the resolver.

This function should be called multiple times. On each call, it will return the next address in connection order preference. If all addresses have been exhausted, this function returns none. After which, you should release the stream with drop-resolve-address-stream.

This function never returns IPv4-mapped IPv6 addresses.

fn drop_resolve_address_stream<'life0, 'async_trait>( &'life0 mut self, this: ResolveAddressStream ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Dispose of the specified resolve-address-stream, after which it may no longer be used.

Note: this function is scheduled to be removed when Resources are natively supported in Wit.

fn non_blocking<'life0, 'async_trait>( &'life0 mut self, this: ResolveAddressStream ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get/set the blocking mode of the stream.

By default a stream is in “blocking” mode, meaning that any function blocks and waits for its completion. When switched to “non-blocking” mode, operations that would block return an again error. After which the API consumer is expected to call subscribe and wait for completion using the wasi-poll module.

Note: these functions are here for WASI Preview2 only. They’re planned to be removed when future is natively supported in Preview3.

fn set_non_blocking<'life0, 'async_trait>( &'life0 mut self, this: ResolveAddressStream, value: bool ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn subscribe<'life0, 'async_trait>( &'life0 mut self, this: ResolveAddressStream ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a pollable which will resolve once the stream is ready for I/O.

Note: this function is here for WASI Preview2 only. It’s planned to be removed when future is natively supported in Preview3.