Struct spin_core::Data

pub struct Data<T> { /* private fields */ }

Host state data associated with individual Stores and Instances.

Implementations

impl<T> Data<T>

pub fn memory_consumed(&self) -> u64

Get the amount of memory in bytes consumed by instances in the store