spin_core
Struct Instance

pub struct Instance(/* private fields */);

An instantiated component.

This type represents an instantiated Component. Instances have exports which can be accessed through functions such as Instance::get_func or Instance::get_export. Instances are owned by a Store and all methods require a handle to the store.

Component instances are created through Linker::instantiate and its family of methods.

This type is similar to the core wasm version wasmtime::Instance except that it represents an instantiated component instead of an instantiated module.

Implementations

impl Instance

pub fn get_func( &self, store: impl AsContextMut, name: impl InstanceExportLookup, ) -> Option<Func>

Looks up an exported function by name within this Instance.

The store argument provided must be the store that this instance lives within and the name argument is the lookup key by which to find the exported function. If the function is found then Some is returned and otherwise None is returned.

The name here can be a string such as &str or it can be a ComponentExportIndex which is loaded prior from a Component.

Panics

Panics if store does not own this instance.

Examples

Looking up a function which is exported from the root of a component:

use wasmtime::{Engine, Store};
use wasmtime::component::{Component, Linker};

let engine = Engine::default();
let component = Component::new(
    &engine,
    r#"
        (component
            (core module $m
                (func (export "f"))
            )
            (core instance $i (instantiate $m))
            (func (export "f")
                (canon lift (core func $i "f")))
        )
    "#,
)?;

// Look up the function by name
let mut store = Store::new(&engine, ());
let instance = Linker::new(&engine).instantiate(&mut store, &component)?;
let func = instance.get_func(&mut store, "f").unwrap();

// The function can also be looked up by an index via a precomputed index.
let (_, export) = component.export_index(None, "f").unwrap();
let func = instance.get_func(&mut store, &export).unwrap();

Looking up a function which is exported from a nested instance:

use wasmtime::{Engine, Store};
use wasmtime::component::{Component, Linker};

let engine = Engine::default();
let component = Component::new(
    &engine,
    r#"
        (component
            (core module $m
                (func (export "f"))
            )
            (core instance $i (instantiate $m))
            (func $f
                (canon lift (core func $i "f")))

            (instance $i
                (export "f" (func $f)))
            (export "i" (instance $i))
        )
    "#,
)?;

// First look up the exported instance, then use that to lookup the
// exported function.
let (_, instance_index) = component.export_index(None, "i").unwrap();
let (_, func_index) = component.export_index(Some(&instance_index), "f").unwrap();

// Then use `func_index` at runtime.
let mut store = Store::new(&engine, ());
let instance = Linker::new(&engine).instantiate(&mut store, &component)?;
let func = instance.get_func(&mut store, &func_index).unwrap();

// Alternatively the `instance` can be used directly in conjunction with
// the `get_export` method.
let instance_index = instance.get_export(&mut store, None, "i").unwrap();
let func_index = instance.get_export(&mut store, Some(&instance_index), "f").unwrap();
let func = instance.get_func(&mut store, &func_index).unwrap();

pub fn get_typed_func<Params, Results>( &self, store: impl AsContextMut, name: impl InstanceExportLookup, ) -> Result<TypedFunc<Params, Results>, Error>
where Params: ComponentNamedList + Lower, Results: ComponentNamedList + Lift,

Looks up an exported Func value by name and with its type.

This function is a convenience wrapper over Instance::get_func and Func::typed. For more information see the linked documentation.

Returns an error if name isnโt a function export or if the exportโs type did not match Params or Results

Panics

Panics if store does not own this instance.

pub fn get_module( &self, store: impl AsContextMut, name: impl InstanceExportLookup, ) -> Option<Module>

Looks up an exported module by name within this Instance.

The store argument provided must be the store that this instance lives within and the name argument is the lookup key by which to find the exported module. If the module is found then Some is returned and otherwise None is returned.

The name here can be a string such as &str or it can be a ComponentExportIndex which is loaded prior from a Component.

For some examples see Instance::get_func for loading values from a component.

Panics

Panics if store does not own this instance.

pub fn get_resource( &self, store: impl AsContextMut, name: impl InstanceExportLookup, ) -> Option<ResourceType>

Looks up an exported resource type by name within this Instance.

The store argument provided must be the store that this instance lives within and the name argument is the lookup key by which to find the exported resource. If the resource is found then Some is returned and otherwise None is returned.

The name here can be a string such as &str or it can be a ComponentExportIndex which is loaded prior from a Component.

For some examples see Instance::get_func for loading values from a component.

Panics

Panics if store does not own this instance.

pub fn get_export( &self, store: impl AsContextMut, instance: Option<&ComponentExportIndex>, name: &str, ) -> Option<ComponentExportIndex>

A methods similar to Component::export_index except for this instance.

This method will lookup the name provided within the instance provided and return a ComponentExportIndex which can be used to pass to other get_* functions like Instance::get_func.

Panics

Panics if store does not own this instance.