Struct spin_core::Reactor

pub struct Reactor { /* private fields */ }

Implementations

impl Reactor

pub fn add_to_linker<T, U>( linker: &mut Linker<T>, get: impl Fn(&mut T) -> &mut U + Send + Sync + Copy + 'static ) -> Result<()>where U: Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Host + Send, T: Send,

pub async fn instantiate_async<T: Send>( store: impl AsContextMut<Data = T>, component: &Component, linker: &Linker<T> ) -> Result<(Self, Instance)>

Instantiates the provided module using the specified parameters, wrapping up the result in a structure that translates between wasm and the host.

pub async fn instantiate_pre<T: Send>( store: impl AsContextMut<Data = T>, instance_pre: &InstancePre<T> ) -> Result<(Self, Instance)>

Instantiates a pre-instantiated module using the specified parameters, wrapping up the result in a structure that translates between wasm and the host.

pub fn new(store: impl AsContextMut, instance: &Instance) -> Result<Self>

Low-level creation wrapper for wrapping up the exports of the instance provided in this structure of wasm exports.

This function will extract exports from the instance defined within store and wrap them all up in the returned structure which can be used to interact with the wasm module.

pub fn inbound_http(&self) -> &InboundHttp

pub fn inbound_redis(&self) -> &InboundRedis