Trait spin_core::tcp::Host

pub trait Host: Sized {

    // Required methods
    fn bind<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        network: Network,
        local_address: IpSocketAddress
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn connect<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        network: Network,
        remote_address: IpSocketAddress
    ) -> Pin<Box<dyn Future<Output = Result<Result<(InputStream, OutputStream), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn listen<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        network: Network
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn accept<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<(TcpSocket, InputStream, OutputStream), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn local_address<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remote_address<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn address_family<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<IpAddressFamily, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn ipv6_only<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_ipv6_only<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_listen_backlog_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn keep_alive<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_keep_alive<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn no_delay<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_no_delay<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unicast_hop_limit<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<u8, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_unicast_hop_limit<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: u8
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn receive_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_receive_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_send_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn non_blocking<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_non_blocking<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        value: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shutdown<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket,
        shutdown_type: ShutdownType
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop_tcp_socket<'life0, 'async_trait>(
        &'life0 mut self,
        this: TcpSocket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Required Methods

fn bind<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, network: Network, local_address: IpSocketAddress ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Bind the socket to a specific network on the provided IP address and port.

If the IP address is zero (0.0.0.0 in IPv4, :: in IPv6), it is left to the implementation to decide which network interface(s) to bind to. If the TCP/UDP port is zero, the socket will be bound to a random free port.

When a socket is not explicitly bound, the first invocation to a listen or connect operation will implicitly bind the socket.

Fails when:

the socket is already bound.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/bind.html

https://man7.org/linux/man-pages/man2/bind.2.html

fn connect<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, network: Network, remote_address: IpSocketAddress ) -> Pin<Box<dyn Future<Output = Result<Result<(InputStream, OutputStream), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Connect to a remote endpoint.

On success:

the socket is transitioned into the Connection state

a pair of streams is returned that can be used to read & write to the connection

Fails when:

the socket is already bound to a different network.

the provided network does not allow connections to the specified endpoint.

the socket is already in the Connection or Listener state.

either the remote IP address or port is 0.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/connect.html

https://man7.org/linux/man-pages/man2/connect.2.html

fn listen<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, network: Network ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Start listening for new connections.

Transitions the socket into the Listener state.

Fails when:

the socket is already bound to a different network.

the provided network does not allow listening on the specified address.

the socket is already in the Connection or Listener state.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/listen.html

https://man7.org/linux/man-pages/man2/listen.2.html

fn accept<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<(TcpSocket, InputStream, OutputStream), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Accept a new client socket.

The returned socket is bound and in the Connection state.

On success, this function returns the newly accepted client socket along with a pair of streams that can be used to read & write to the connection.

Fails when this socket is not in the Listening state.

References:

https://pubs.opengroup.org/onlinepubs/9699919799/functions/accept.html

https://man7.org/linux/man-pages/man2/accept.2.html

fn local_address<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the bound local address.

Returns an error if the socket is not bound.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getsockname.html

https://man7.org/linux/man-pages/man2/getsockname.2.html

fn remote_address<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the bound remote address.

Fails when the socket is not in the Connection state.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getpeername.html

https://man7.org/linux/man-pages/man2/getpeername.2.html

fn address_family<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<IpAddressFamily, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Whether this is a IPv4 or IPv6 socket.

Equivalent to the SO_DOMAIN socket option.

fn ipv6_only<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Whether IPv4 compatibility (dual-stack) mode is disabled or not. Implementations are not required to support dual-stack mode. Calling set-ipv6-only(false) might fail.

Fails when called on an IPv4 socket.

Equivalent to the IPV6_V6ONLY socket option.

fn set_ipv6_only<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: bool ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn set_listen_backlog_size<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Hints the desired listen queue size. Implementations are free to ignore this.

fn keep_alive<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Equivalent to the SO_KEEPALIVE socket option.

fn set_keep_alive<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: bool ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn no_delay<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Equivalent to the TCP_NODELAY socket option.

fn set_no_delay<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: bool ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn unicast_hop_limit<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<u8, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Equivalent to the IP_TTL & IPV6_UNICAST_HOPS socket options.

fn set_unicast_hop_limit<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: u8 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn receive_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

The kernel buffer space reserved for sends/receives on this socket.

Note #1: an implementation may choose to cap or round the buffer size when setting the value. In other words, after setting a value, reading the same setting back may return a different value.

Note #2: there is not necessarily a direct relationship between the kernel buffer size and the bytes of actual data to be sent/received by the application, because the kernel might also use the buffer space for internal metadata structures.

Fails when this socket is in the Listening state.

Equivalent to the SO_RCVBUF and SO_SNDBUF socket options.

fn set_receive_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn send_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn set_send_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn non_blocking<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get/set the blocking mode of the socket.

By default a socket is in “blocking” mode, meaning that any function blocks and waits for its completion. When switched to “non-blocking” mode, operations that would block return an again error. After which the API consumer is expected to call subscribe and wait for completion using the wasi-poll module.

Note: these functions are here for WASI Preview2 only. They’re planned to be removed when future is natively supported in Preview3.

fn set_non_blocking<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, value: bool ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn subscribe<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a pollable which will resolve once the socket is ready for I/O.

Note: this function is here for WASI Preview2 only. It’s planned to be removed when future is natively supported in Preview3.

fn shutdown<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket, shutdown_type: ShutdownType ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Gracefully shut down the connection.

receive: the socket is not expecting to receive any more data from the peer. All subsequent read operations on the input-stream associated with this socket will return an End Of Stream indication. Any data still in the receive queue at time of calling shutdown will be discarded.

send: the socket is not expecting to send any more data to the peer. All subsequent write operations on the output-stream associated with this socket will return an error.

both: same effect as receive & send combined.

The shutdown function does not close the socket.

Fails when the socket is not in the Connection state.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/shutdown.html

https://man7.org/linux/man-pages/man2/shutdown.2.html

fn drop_tcp_socket<'life0, 'async_trait>( &'life0 mut self, this: TcpSocket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Dispose of the specified tcp-socket, after which it may no longer be used.

Note: this function is scheduled to be removed when Resources are natively supported in Wit.