Trait spin_core::udp::Host

pub trait Host: Sized {

    // Required methods
    fn bind<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        network: Network,
        local_address: IpSocketAddress
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn connect<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        network: Network,
        remote_address: IpSocketAddress
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn receive<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<Datagram, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        datagram: Datagram
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn local_address<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remote_address<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn address_family<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<IpAddressFamily, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn ipv6_only<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_ipv6_only<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        value: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unicast_hop_limit<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<u8, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_unicast_hop_limit<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        value: u8
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn receive_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_receive_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        value: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn send_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_send_buffer_size<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        value: u64
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn non_blocking<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_non_blocking<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket,
        value: bool
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop_udp_socket<'life0, 'async_trait>(
        &'life0 mut self,
        this: UdpSocket
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Required Methods

fn bind<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, network: Network, local_address: IpSocketAddress ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Bind the socket to a specific network on the provided IP address and port.

If the IP address is zero (0.0.0.0 in IPv4, :: in IPv6), it is left to the implementation to decide which network interface(s) to bind to. If the TCP/UDP port is zero, the socket will be bound to a random free port.

When a socket is not explicitly bound, the first invocation to a connect, send or receive operation will implicitly bind the socket.

Fails when:

the socket is already bound.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/bind.html

https://man7.org/linux/man-pages/man2/bind.2.html

fn connect<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, network: Network, remote_address: IpSocketAddress ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Set the destination address.

The local-address is updated based on the best network path to remote-address.

When a destination address is set:

all receive operations will only return datagrams sent from the provided remote-address.

the send function can only be used to send to this destination.

Note that this function does not generate any network traffic and the peer is not aware of this “connection”.

Fails when:

the socket is already bound to a different network.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/connect.html

https://man7.org/linux/man-pages/man2/connect.2.html

fn receive<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<Datagram, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Receive a message.

Returns:

The sender address of the datagram

The number of bytes read.

Fails when:

the socket is not bound.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/recvfrom.html

https://pubs.opengroup.org/onlinepubs/9699919799/functions/recvmsg.html

https://man7.org/linux/man-pages/man2/recv.2.html

fn send<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, datagram: Datagram ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Send a message to a specific destination address.

The remote address option is required. To send a message to the “connected” peer, call remote-address to get their address.

Fails when:

the socket is not bound. Unlike POSIX, this function does not perform an implicit bind.

the socket is in “connected” mode and the datagram.remote-address does not match the address passed to connect.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/sendto.html

https://pubs.opengroup.org/onlinepubs/9699919799/functions/sendmsg.html

https://man7.org/linux/man-pages/man2/send.2.html

fn local_address<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the current bound address.

Returns an error if the socket is not bound.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getsockname.html

https://man7.org/linux/man-pages/man2/getsockname.2.html

fn remote_address<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<IpSocketAddress, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the address set with connect.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getpeername.html

https://man7.org/linux/man-pages/man2/getpeername.2.html

fn address_family<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<IpAddressFamily, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Whether this is a IPv4 or IPv6 socket.

Equivalent to the SO_DOMAIN socket option.

fn ipv6_only<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Whether IPv4 compatibility (dual-stack) mode is disabled or not. Implementations are not required to support dual-stack mode, so calling set-ipv6-only(false) might fail.

Fails when called on an IPv4 socket.

Equivalent to the IPV6_V6ONLY socket option.

fn set_ipv6_only<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, value: bool ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn unicast_hop_limit<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<u8, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Equivalent to the IP_TTL & IPV6_UNICAST_HOPS socket options.

fn set_unicast_hop_limit<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, value: u8 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn receive_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

The kernel buffer space reserved for sends/receives on this socket.

Note #1: an implementation may choose to cap or round the buffer size when setting the value. In other words, after setting a value, reading the same setting back may return a different value.

Note #2: there is not necessarily a direct relationship between the kernel buffer size and the bytes of actual data to be sent/received by the application, because the kernel might also use the buffer space for internal metadata structures.

Fails when this socket is in the Listening state.

Equivalent to the SO_RCVBUF and SO_SNDBUF socket options.

fn set_receive_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, value: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn send_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<u64, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn set_send_buffer_size<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, value: u64 ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn non_blocking<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get/set the blocking mode of the socket.

By default a socket is in “blocking” mode, meaning that any function blocks and waits for its completion. When switched to “non-blocking” mode, operations that would block return an again error. After which the API consumer is expected to call subscribe and wait for completion using the wasi-poll module.

Note: these functions are here for WASI Preview2 only. They’re planned to be removed when future is natively supported in Preview3.

fn set_non_blocking<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket, value: bool ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn subscribe<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<Pollable>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a pollable which will resolve once the socket is ready for I/O.

Note: this function is here for WASI Preview2 only. It’s planned to be removed when future is natively supported in Preview3.

fn drop_udp_socket<'life0, 'async_trait>( &'life0 mut self, this: UdpSocket ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Dispose of the specified udp-socket, after which it may no longer be used.

Note: this function is scheduled to be removed when Resources are natively supported in Wit.