Crate spin_core

Spin core execution engine

This crate provides low-level Wasm and WASI functionality required by Spin. Most of this functionality consists of wrappers around wasmtime and wasmtime_wasi that narrows the flexibility of wasmtime to the set of features used by Spin (such as only supporting wasmtime’s async calling style).

Re-exports

pub use wasmtime;

Structs

Config

Global configuration for EngineBuilder.

Data

Host state data associated with individual Stores and Instances.

Engine

An Engine is a global context for the initialization and execution of Spin components.

EngineBuilder

A builder interface for configuring a new Engine.

HostComponentDataHandle

An opaque handle returned by crate::EngineBuilder::add_host_component which can be passed to HostComponentsData to access or set associated HostComponent::Data.

HostComponentsData

Holds a heterogenous set of HostComponent::Datas.

I32Exit

An error returned from the proc_exit host syscall.

Instance

An instantiated WebAssembly module.

InstancePre

A pre-initialized instance that is ready to be instantiated.

Module

A compiled WebAssembly module, ready to be instantiated.

OutputBuffer

An in-memory stdio output buffer.

Store

A Store holds the runtime state of a Spin instance.

StoreBuilder

A builder interface for configuring a new Store.

Enums

Trap

Constants

DEFAULT_EPOCH_TICK_INTERVAL

The default EngineBuilder::epoch_tick_interval.

Traits

HostComponent

A trait for Spin “host components”.

Type Definitions

Linker

An alias for wasmtime::Linker specialized to Data.